chatbot/pyproject.toml
[project]
name = "wallet-chatbot"
version = "0.1.0"
description = "AI Agent for E-Wallet system using LangGraph"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]


[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent", "wallet_chatbot"]
[tool.setuptools.package-dir]
"agent" = "src/agent"
"wallet_chatbot" = "src"

